Move src_to_cmds function
Move src_to_cmds function from run.rs to read_src.rs
