+++ src/read_src.rs
@@ -1,4 +1,5 @@
 use std::path::PathBuf;
+use std::collections::HashMap;
 use crate::init;
 
 pub(crate) type Src = Vec<Vec<String>>;
@@ -24,6 +25,37 @@ fn conv_token(src: String) -> Src {
         token.push(tmp);
     }
     token
+}
+
+fn src_to_cmds(src: Src) -> HashMap<String,Src> {
+    let mut cmds: HashMap<String,Src> = HashMap::new();
+
+    let mut key = String::new();
+    let mut val = Vec::new();
+    let mut fn_flag = false;
+    for lines in src {
+        let mut tmp = Vec::new();
+        for cmd in lines {
+            if key == String::new() {
+                key = cmd.clone();
+            } else if cmd == "{".to_string() {
+                fn_flag = true;
+            } else if cmd == "}".to_string() {
+                fn_flag = false;
+            } else {
+                tmp.push(cmd);
+            }
+        }
+        if tmp.len() > 0 {
+            val.push(tmp.clone());
+        }
+        if !fn_flag & (key != "".to_string()) {
+            cmds.insert(key.clone(),val.clone());
+            key = String::new();
+            val = Vec::new();
+        }
+    }
+    cmds
 }
 
 pub(crate) fn get_src(src: init::EnvConf) -> Src {
